# swagger/generated-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Telegram Hisobchi Bot API",
    "version": "2.0.0",
    "description": "API documentation for Telegram Hisobchi Bot Backend - Advanced Sales Management System with SellerStock Management",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "TelegramAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-telegram-id",
        "description": "Telegram user ID (REQUIRED for all authenticated endpoints)"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "telegramId": {
            "type": "string",
            "description": "Telegram user ID"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "firstName": {
            "type": "string",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "description": "Last name"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "seller"],
            "description": "User role"
          },
          "assignedProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "description": "Products assigned to seller"
          },
          "isActive": {
            "type": "boolean",
            "description": "User active status"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Category ID"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Category creation date"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Product ID"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "description": "Selling price"
          },
          "costPrice": {
            "type": "number",
            "description": "Cost price"
          },
          "category": {
            "type": "object",
            "ref": "#/components/schemas/Category",
            "description": "Product category"
          },
          "count": {
            "type": "number",
            "description": "Warehouse stock count"
          },
          "image": {
            "type": "string",
            "description": "Product image URL"
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit"
          },
          "color": {
            "type": "string",
            "description": "Product color"
          },
          "isActive": {
            "type": "boolean",
            "description": "Product active status"
          },
          "assignedSellers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sellers assigned to this product"
          }
        }
      },
      "SellerStock": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "SellerStock ID"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "number",
            "description": "Stock quantity with seller",
            "minimum": 0
          },
          "lastTransferDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date of last transfer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update date"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "TelegramAuth": []
    }
  ],
  "paths": {
    "/api/admin/sellers": {
      "get": {
        "tags": ["Admin - User Management"],
        "summary": "Get all sellers",
        "description": "Returns a list of all sellers",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sellers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sellers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/admin/sellers/{sellerId}/products/{productId}/assign": {
      "post": {
        "tags": ["Admin - Product Assignment"],
        "summary": "Assign product to seller",
        "description": "Assigns a product to a seller with optional quantity transfer",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seller ID"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Quantity to transfer (0 for permission only)",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Product assigned successfully"
                    },
                    "seller": {
                      "$ref": "#/components/schemas/User"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Insufficient warehouse stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Seller or product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/sellers/{sellerId}/products/{productId}/unassign": {
      "delete": {
        "tags": ["Admin - Product Assignment"],
        "summary": "Unassign product from seller",
        "description": "Removes product assignment from seller with stock handling",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seller ID"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          },
          {
            "in": "query",
            "name": "returnStock",
            "schema": {
              "type": "boolean"
            },
            "description": "Whether to return seller's stock to warehouse",
            "example": true
          }
        ],
        "responses": {
          "200": {
            "description": "Product unassigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Product unassigned successfully"
                    },
                    "stockReturned": {
                      "type": "number",
                      "description": "Amount of stock returned to warehouse",
                      "example": 15
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Seller has stock, use returnStock=true",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Cannot unassign. Seller has 15 units in stock. Return stock first or use ?returnStock=true"
                    },
                    "currentStock": {
                      "type": "number",
                      "example": 15
                    },
                    "suggestion": {
                      "type": "string",
                      "example": "Use: DELETE /admin/sellers/12345/products/67890?returnStock=true"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Seller or product not found"
          }
        }
      }
    },
    "/api/admin/seller-stocks": {
      "get": {
        "tags": ["Admin - Stock Management"],
        "summary": "Get all seller stocks",
        "description": "Returns a list of all seller stock records",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Seller stocks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sellerStocks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SellerStock"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/api/admin/seller-stocks/{stockId}": {
      "patch": {
        "tags": ["Admin - Stock Management"],
        "summary": "Update seller stock quantity",
        "description": "Updates seller stock quantity with warehouse adjustment and audit trail",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "stockId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "SellerStock ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quantity"],
                "properties": {
                  "quantity": {
                    "type": "number",
                    "minimum": 0,
                    "description": "New stock quantity",
                    "example": 25
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock quantity updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Stock quantity updated successfully"
                    },
                    "sellerStock": {
                      "$ref": "#/components/schemas/SellerStock"
                    },
                    "change": {
                      "type": "number",
                      "description": "Quantity change (positive = increase, negative = decrease)",
                      "example": 5
                    },
                    "warehouseStockAfter": {
                      "type": "number",
                      "description": "Warehouse stock after adjustment",
                      "example": 45
                    },
                    "transferCreated": {
                      "type": "boolean",
                      "description": "Whether a transfer record was created",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid quantity or insufficient warehouse stock",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Cannot increase stock by 10. Warehouse only has 5 units available."
                    },
                    "warehouseStock": {
                      "type": "number",
                      "example": 5
                    },
                    "requestedIncrease": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "SellerStock not found"
          }
        }
      },
      "delete": {
        "tags": ["Admin - Stock Management"],
        "summary": "Delete seller stock",
        "description": "Removes seller stock record and returns quantity to warehouse",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "stockId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "SellerStock ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Seller stock removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Seller stock removed successfully"
                    },
                    "returnedToWarehouse": {
                      "type": "number",
                      "description": "Quantity returned to warehouse",
                      "example": 20
                    },
                    "sellerStock": {
                      "$ref": "#/components/schemas/SellerStock"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "SellerStock not found"
          }
        }
      }
    },
    "/api/admin/sellers/{sellerId}/stocks": {
      "get": {
        "tags": ["Admin - Stock Management"],
        "summary": "Get stocks for specific seller",
        "description": "Returns all stock records for a specific seller",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sellerId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seller ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Seller stocks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sellerStocks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SellerStock"
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "Message if no stocks found",
                      "example": "No stocks found for this seller"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{productId}/stocks": {
      "get": {
        "tags": ["Admin - Stock Management"],
        "summary": "Get stocks for specific product",
        "description": "Returns all seller stock records for a specific product",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product stocks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "productStocks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SellerStock"
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "Message if no stocks found",
                      "example": "No stocks found for this product"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/reports": {
      "get": {
        "tags": ["Admin - Reports"],
        "summary": "Get comprehensive report",
        "description": "Returns detailed sales report with analytics",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "schema": {
              "type": "string"
            },
            "description": "Report start date (defaults to current date)",
            "example": "2024-01-01"
          },
          {
            "in": "query",
            "name": "end",
            "schema": {
              "type": "string"
            },
            "description": "Report end date (defaults to current date)",
            "example": "2024-01-31"
          }
        ],
        "responses": {
          "200": {
            "description": "Range report retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}